moved `Node` related code to `node` package
