+++ compiler-access/src/main/scala/org/alephium/ralph/lsp/access/compiler/ast/node/RootPosition.scala
@@ -1,4 +1,4 @@
-package org.alephium.ralph.lsp.access.compiler.ast
+package org.alephium.ralph.lsp.access.compiler.ast.node
 
 import org.alephium.ralph.Ast.Positioned
 import org.alephium.ralph.SourceIndex
