+++ compiler-access/src/test/scala/org/alephium/ralph/lsp/access/compiler/ast/NodeWalkDownSpec.scala
@@ -1,5 +1,6 @@
 package org.alephium.ralph.lsp.access.compiler.ast
 
+import org.alephium.ralph.lsp.access.compiler.ast.node.Node
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
